moved DiagnosticStatusPanel to the proper dir
