+++ packages/suite-base/src/panels/DiagnosticStatus/DiagnosticStatusPanel.stories.tsx
@@ -7,13 +7,13 @@
 
 import { StoryObj } from "@storybook/react";
 
-import DiagnosticStatusPanel from "@lichtblick/suite-base/panels/diagnostics/DiagnosticStatusPanel";
+import DiagnosticStatusPanel from "@lichtblick/suite-base/panels/DiagnosticStatus";
 import { makeDiagnosticMessage } from "@lichtblick/suite-base/panels/diagnostics/DiagnosticSummary.stories";
 import { LEVELS } from "@lichtblick/suite-base/panels/diagnostics/constants";
 import PanelSetup, { Fixture } from "@lichtblick/suite-base/stories/PanelSetup";
 
 export default {
-  title: "panels/diagnostics/DiagnosticStatusPanel",
+  title: "panels/DiagnosticStatus/DiagnosticStatusPanel",
 };
 
 const fixture: Fixture = {
