+++ packages/suite-base/src/panels/DiagnosticStatus/DiagnosticStatusPanel.style.ts
