+++ packages/suite-base/src/panels/DiagnosticStatus/DiagnosticTable.style.ts
