+++ packages/suite-base/src/panels/DiagnosticStatus/assets/diagnostic-status.png
