+++ packages/suite-base/src/panels/DiagnosticStatus/hooks/useAvailableDiagnostics.ts
@@ -8,7 +8,7 @@
 import { useMemo } from "react";
 
 import { useMessageReducer } from "@lichtblick/suite-base/PanelAPI";
-import { DiagnosticStatusArrayMsg } from "@lichtblick/suite-base/panels/diagnostics/types";
+import { DiagnosticStatusArrayMsg } from "@lichtblick/suite-base/panels/DiagnosticStatus/types";
 import { MessageEvent } from "@lichtblick/suite-base/players/types";
 import { mightActuallyBePartial } from "@lichtblick/suite-base/util/mightActuallyBePartial";
