+++ packages/suite-base/src/panels/DiagnosticStatus/index.tsx
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+import Panel from "@lichtblick/suite-base/components/Panel";
+import DiagnosticStatusPanel from "@lichtblick/suite-base/panels/DiagnosticStatus/DiagnosticStatusPanel";
+import { DEFAULT_CONFIG } from "@lichtblick/suite-base/panels/DiagnosticStatus/constants";
+
+export default Panel(
+  Object.assign(DiagnosticStatusPanel, {
+    panelType: "DiagnosticStatusPanel",
+    defaultConfig: DEFAULT_CONFIG,
+  }),
+);
