+++ packages/suite-base/src/panels/DiagnosticStatus/utils/sanitize.ts
@@ -1,0 +1,18 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+import sanitizeHtml from "sanitize-html";
+
+import { ALLOWED_TAGS } from "@lichtblick/suite-base/panels/DiagnosticStatus/constants";
+
+export function sanitize(value: string): { __html: string } {
+  return {
+    __html: sanitizeHtml(value, {
+      allowedTags: ALLOWED_TAGS,
+      allowedAttributes: {
+        font: ["color", "size"],
+        td: ["colspan"],
+        th: ["colspan"],
+      },
+    }),
+  };
+}
