+++ packages/suite-base/src/panels/diagnostics/constants.ts
@@ -1,10 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
 // SPDX-License-Identifier: MPL-2.0
 
-import {
-  DiagnosticStatusConfig,
-  DiagnosticSummaryConfig,
-} from "@lichtblick/suite-base/panels/diagnostics/types";
+import { DiagnosticSummaryConfig } from "@lichtblick/suite-base/panels/diagnostics/types";
 
 // Trim the message if it's too long. We sometimes get crazy massive messages here that can
 // otherwise crash our entire UI. I looked at a bunch of messages manually and they are typically
@@ -43,13 +40,6 @@ export const MESSAGE_COLORS: Record<number, string> = {
   [LEVELS.STALE]: "text.secondary",
 };
 
-// export const MESSAGE_COLORS: { [key: string]: string } = {
-//   ok: "success.main",
-//   warn: "warning.main",
-//   error: "error.main",
-//   stale: "text.secondary",
-// };
-
 export const diagnosticSummaryConfig: DiagnosticSummaryConfig = {
   minLevel: 0,
   pinnedIds: [],
@@ -57,5 +47,3 @@ export const diagnosticSummaryConfig: DiagnosticSummaryConfig = {
   topicToRender: "/diagnostics",
   sortByLevel: true,
 };
-
-export const diagnosticStatusConfig: DiagnosticStatusConfig = { topicToRender: "/diagnostics" };
