+++ packages/suite-base/src/panels/diagnostics/hooks/useDiagnostics.test.ts
@@ -17,8 +17,8 @@
 import { renderHook } from "@testing-library/react";
 
 import * as PanelAPI from "@lichtblick/suite-base/PanelAPI";
+import { DiagnosticStatusArrayMsg } from "@lichtblick/suite-base/panels/DiagnosticStatus/types";
 import { LEVELS } from "@lichtblick/suite-base/panels/diagnostics/constants";
-import { DiagnosticStatusArrayMsg } from "@lichtblick/suite-base/panels/diagnostics/types";
 import BasicBuilder from "@lichtblick/suite-base/testing/builders/BasicBuilder";
 import DiagnosticsBuilder from "@lichtblick/suite-base/testing/builders/DiagnosticsBuilder";
 import MessageEventBuilder from "@lichtblick/suite-base/testing/builders/MessageEventBuilder";
