+++ packages/suite-base/src/testing/builders/DiagnosticsBuilder.ts
@@ -6,9 +6,9 @@ import {
   DiagnosticStatusArrayMsg,
   DiagnosticStatusConfig,
   DiagnosticStatusMessage,
-  DiagnosticSummaryConfig,
   KeyValue,
-} from "@lichtblick/suite-base/panels/diagnostics/types";
+} from "@lichtblick/suite-base/panels/DiagnosticStatus/types";
+import { DiagnosticSummaryConfig } from "@lichtblick/suite-base/panels/diagnostics/types";
 import BasicBuilder from "@lichtblick/suite-base/testing/builders/BasicBuilder";
 import RosTimeBuilder from "@lichtblick/suite-base/testing/builders/RosTimeBuilder";
 import { defaults } from "@lichtblick/suite-base/testing/builders/utilities";
